+++ python_search/ps_llm/tasks/base_task.py
@@ -7,8 +7,6 @@ class BaseDataTask:
     def build_dataset(self):
         raise NotImplementedError()
 
-
-
     def predict(self, content):
         from python_search.ps_llm.t5.t5_model import T5Model
         model = T5Model.load_productionalized_model()