+++ python_search/search_ui/fzf.py
@@ -41,7 +41,6 @@ class Fzf:
         --margin=0% \
         --padding=0% \
         {self._run_key("enter")} \
-        {self._run_key("ctrl-t", wrap_in_terminal=True)} \
         {self._run_key("double-click")} \
         {self._edit_key('ctrl-e')} \
         {self._edit_key('right-click')} \
@@ -54,6 +53,7 @@ class Fzf:
         --bind "ctrl-c:execute-silent:(nohup python_search _copy_entry_content {{}})" \
         --bind "ctrl-n:execute-silent:(nohup register_new launch_from_fzf {{}} & )" \
         --bind "ctrl-g:execute-silent:(google_it search {{q}})" \
+        --bind "ctrl-d:execute-silent:(monorepo_cli  register_german {{q}} &)" \
         --bind "ctrl-y:execute-silent:(python_search _copy_key_only {{}})" \
         --bind "ctrl-r:reload-sync:(ps_search --reload_enabled=True)" \
         --bind "ctrl-n:reload-sync:(ps_search)" \
@@ -63,7 +63,6 @@ class Fzf:
         --bind "esc:execute-silent:(ps_fzf hide_current_focused_window)" \
         --bind "esc:+clear-query" \
         --bind "ctrl-k:abort" \
-        --bind "ctrl-d:abort"  \
         {self._get_fzf_theme()} ; exit 0
         '
         """
@@ -87,8 +86,7 @@ class Fzf:
 
         return f"""--bind "{shortcut}:execute-silent:(nohup run_key {{}}  --query_used {{q}} {wrap_in_terminal_expr} {{}} &)" \
         --bind "{shortcut}:+reload-sync:(sleep 3 && ps_search)" \
-        --bind "{shortcut}:+first" \
-        --bind "{shortcut}:+clear-screen" """
+        --bind "{shortcut}:+first" """
 
     def _edit_key(self, shortcut) -> str:
         return f''' --bind "{shortcut}:execute-silent:(entries_editor edit_key {{}} & disown )" \
@@ -97,10 +95,80 @@ class Fzf:
         if self.configuration.get_fzf_theme() == "light":
             return "--color=bg+:#ffffff,bg:#ffffff,hl:#719872,fg:#616161,header:#719872,info:#727100,pointer:#E12672,marker:#E17899,fg+:#616161,preview-bg:#ffffff,prompt:#0099BD,hl+:#719899"
 
+
+        if self.configuration.get_fzf_theme() in ["solarized"]:
+            theme = Theme().solarized()
+            return f' --color="fg:{theme.fg},bg:{theme.bg},hl:{theme.hl},fg+:{theme.fg_plus},bg+:{theme.bg_plus},hl+:{theme.hl_plus},info:{theme.info},prompt:{theme.prompt_arrows},pointer:{theme.pointer_current_line},marker:{theme.marker},spinner:{theme.spinner},header:{theme.header},query:{theme.query},label:{theme.label},border:{theme.border}" '
+
         if self.configuration.get_fzf_theme() in ["dark", "dracula"]:
-            return ' --color="fg:#f8f8f2,bg:#282a36,hl:#bd93f9,fg+:#f8f8f2,bg+:#44475a,hl+:#bd93f9,info:#ffb86c,prompt:#50fa7b,pointer:#ff79c6,marker:#ff79c6,spinner:#ffb86c,header:#6272a4" '
+            return ' --color="fg:#f8f8f2,bg:#282a36,hl:#bd93f9,fg+:#f8f8f2,bg+:#44475a,hl+:#bd93f9,info:#ffb86c,prompt:#50fa7b,pointer:#ff79c6,marker:#ff79c6,spinner:#ffb86c,header:#586e75" '
 
         return " "
+
+
+class Theme:
+    """
+       fg                  Text
+         preview-fg        Preview window text
+       bg                  Background
+         preview-bg        Preview window background
+       hl                  Highlighted substrings
+       fg+                 Text (current line)
+       bg+                 Background (current line)
+         gutter            Gutter on the left
+       hl+                 Highlighted substrings (current line)
+       query               Query string
+         disabled          Query string when search is disabled (--disabled)
+       info                Info line (match counters)
+       border              Border around the window (--border and --preview)
+         scrollbar         Scrollbar
+         preview-border    Border around the preview window (--preview)
+         preview-scrollbar Scrollbar
+         separator         Horizontal separator on info line
+       label               Border label (--border-label and --preview-label)
+         preview-label     Border label of the preview window (--preview-label)
+       prompt              Prompt
+       pointer             Pointer to the current line
+       marker              Multi-select marker
+       spinner             Streaming input indicator
+       header              Header
+    """
+    fg: str
+    bg: str
+    hl: str
+    fg_plus: str
+    bg_plus: str
+    hl_plus: str
+    info: str
+    prompt_arrows: str
+    # the pointer of hte current line
+    pointer_current_line: str
+    marker: str
+    spinner: str
+    header: str
+    query: str
+
+    def solarized(self):
+        self.fg = "#2aa198"
+        self.bg = "#002b36"
+        self.hl = "#859900"
+        self.fg_plus = "#6c71c4"
+        self.bg_plus = "#002b36"
+        self.hl_plus = "#268bd2"
+        self.info = "#d33682"
+        self.prompt_arrows = "#002b36"
+
+        self.pointer_current_line = "#6c71c4"
+        self.marker = "#268bd2"
+        self.spinner = "#2aa198"
+        self.header = "#268bd2"
+        self.query = "#cb4b16"
+        self.label = self.bg
+        self.border = "#073642"
+
+        return self
+
+